Revert "Basics for piFpaF (no smoke)"
This reverts commit a4a4f5f42c4aace1ebb57db1137560a11b14c3fb.

+++ OpenGL/src/Basics/Coords.h
@@ -14,7 +14,7 @@ struct Coords
 public:
 	double posX;
 	double posY;
-	double Velocity;
+	float Velocity;
 };
 
 #endif
